+++ config.h
@@ -75,12 +75,15 @@ struct config {
     struct {
         enum cursor_style style;
         bool blink;
-        bool hide_when_typing;
         struct {
             uint32_t text;
             uint32_t cursor;
         } color;
     } cursor;
+
+    struct {
+        bool hide_when_typing;
+    } mouse;
 
     struct {
         /* Bindings for "normal" mode */